+++ packages/rocketh-diamond/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocketh/diamond
 
+## 0.19.2
+
+### Patch Changes
+
+- export artifacts files
+
 ## 0.19.1
 
 ### Patch Changes